+++ widgets/secrets/src/widget.utils.ts
@@ -1,0 +1,1 @@
+export const translateSecrets = Stage.Utils.getT('widgets.secrets');
